made structure for navigation post logout, modified dashboard, and changed theme
