+++ frontend/src/components/signup/Signup.jsx
@@ -1,5 +1,5 @@
 // import React from 'react'
-import React, { useContext, useState } from 'react';
+import { useContext, useState } from 'react';
 import { UserContext } from '../../userContext'
 import { GoEye, GoEyeClosed } from "react-icons/go";
 import { Link, useNavigate } from "react-router-dom";
